+++ packages/ts-newskit/src/components/misc/toast/__tests__/toast.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render } from '../../../utils/test-utils';
+import { render } from '../../../../utils/test-utils';
 
 import { Toast } from '../toast';