+++ packages/ts-newskit/src/components/navigation/global-nav/topnav/__tests__/topnav.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { fireEvent } from '@testing-library/react';
-import { render, screen } from '../../../../utils/test-utils';
+import { render, screen } from '../../../../../utils/test-utils';
 import '@testing-library/jest-dom';
 import { TopNav } from '../topnav';
 import NavSearch from '../../search';