+++ packages/ts-newskit/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.28.0](https://github.com/newsuk/times-components/compare/@times-components/ts-newskit@1.27.0...@times-components/ts-newskit@1.28.0) (2023-04-27)
+
+
+### Features
+
+* **TMRX-1198:** in article audio player ([#3336](https://github.com/newsuk/times-components/issues/3336)) ([ab52180](https://github.com/newsuk/times-components/commit/ab521804abad058f6d0ebd97736b9b1d26a06203))
+
+
+
+
+
+# [1.27.0](https://github.com/newsuk/times-components/compare/@times-components/ts-newskit@1.26.0...@times-components/ts-newskit@1.27.0) (2023-04-27)
+
+
+### Features
+
+* **TMST-618:** Breakout section component ([#3321](https://github.com/newsuk/times-components/issues/3321)) ([58d60ff](https://github.com/newsuk/times-components/commit/58d60ffd9f3a94c208c682ac0536d62c9bdce4be))
+
+
+
+
+
+# [1.26.0](https://github.com/newsuk/times-components/compare/@times-components/ts-newskit@1.25.0...@times-components/ts-newskit@1.26.0) (2023-04-27)
+
+
+### Features
+
+* **TMRX-1199:** create feedback component ([#3332](https://github.com/newsuk/times-components/issues/3332)) ([0604aff](https://github.com/newsuk/times-components/commit/0604affc94d87767dfa56cda5887885d3c242d7e))
+
+
+
+
+
 # [1.25.0](https://github.com/newsuk/times-components/compare/@times-components/ts-newskit@1.24.0...@times-components/ts-newskit@1.25.0) (2023-04-24)